\name{additional.linear.constraints}
\alias{additional.linear.constraints}
\title{Additional Linear Constraints}
\description{internal function for sisus}
\usage{additional.linear.constraints(constraints.cols, constraints.type, 
    constraints.RHS.b, constraints.sources, n.sources, lc.include.sw)}
\arguments{
  \item{constraints.cols}{internal variable}
  \item{constraints.type}{internal variable}
  \item{constraints.RHS.b}{internal variable}
  \item{constraints.sources}{internal variable}
  \item{n.sources}{internal variable}
  \item{lc.include.sw}{internal variable}
}
\details{interal function for sisus.run()}
\value{internal variable}

\author{Erik Barry Erhardt <erik@statacumen.com> [aut, cre]}





