\name{assign.variables}
\alias{assign.variables}
\title{Assign variables from input file to variable names}
\description{internal function for sisus}
\usage{assign.variables(DATA, first = 0)}
\arguments{
  \item{DATA}{internal variable}
  \item{first}{internal variable}
}
\details{interal function for sisus.run()}
\value{internal variable}

\author{Erik Barry Erhardt <erik@statacumen.com> [aut, cre]}





