\name{diag.panel.hist}
\alias{diag.panel.hist}
\title{define scatterplot diagonal formatting}
\description{internal function for sisus}
\usage{diag.panel.hist(x, ...)}
\arguments{
  \item{x}{internal variable}
  \item{\dots}{internal variable}
}
\details{interal function for sisus.run()}
\value{internal variable}

\author{Erik Barry Erhardt <erik@statacumen.com>}





