\name{dirichlet.moments}
\alias{dirichlet.moments}
\title{Calculate the moments of the Dirichlet distribution given the values of vector alpha}
\description{internal function for sisus}
\usage{dirichlet.moments(alpha.in, priors.precision, priors.sources)}
\arguments{
  \item{alpha.in}{internal variable}
  \item{priors.precision}{internal variable}
  \item{priors.sources}{internal variable}
}
\details{interal function for sisus.run()}
\value{internal variable}

\author{Erik Barry Erhardt <erik@statacumen.com>}





