\name{filename.clean}
\alias{filename.clean}
\title{Clean up filenames, replacing spaces and bad characters with "_" and "-"}
\description{internal function for sisus}
\usage{filename.clean(filename)}
\arguments{
  \item{filename}{internal variable}
}
\details{interal function for sisus.run()}
\value{internal variable}

\author{Erik Barry Erhardt <erik@statacumen.com> [aut, cre]}





