\name{get.data}
\alias{get.data}
\title{Get data from Excel spreadsheet}
\description{internal function for sisus}
\usage{get.data(filename)}
\arguments{
  \item{filename}{internal variable}
}
\details{interal function for sisus.run()}
\value{internal variable}

\author{Erik Barry Erhardt <erik@statacumen.com>}





