\name{indy.mixture.isotope.mvn.sample}
\alias{indy.mixture.isotope.mvn.sample}
\title{Strip out the samples associated with the current mixture}
\description{internal function for sisus}
\usage{indy.mixture.isotope.mvn.sample(isotope.mvn.sample, i.mixture, 
    n.mixtures, n.sources, n.isotopes)}
\arguments{
  \item{isotope.mvn.sample}{internal variable}
  \item{i.mixture}{internal variable}
  \item{n.mixtures}{internal variable}
  \item{n.sources}{internal variable}
  \item{n.isotopes}{internal variable}
}
\details{interal function for sisus.run()}
\value{internal variable}

\author{Erik Barry Erhardt <erik@statacumen.com> [aut, cre]}





