\name{mcmc.diagnostics}
\alias{mcmc.diagnostics}
\title{MCMC Diagnostics}
\description{internal function for sisus}
\usage{mcmc.diagnostics(p.biomass.sam, p.results.name, names.mixtures.indy, 
    n.sources, names.sources, n.isotopes, names.isotopes, M.actual, 
    output.mcmc.diagnostics.filename, filename.prefix, analysis.name, 
    plot.format.list)}
\arguments{
  \item{p.biomass.sam}{internal variable}
  \item{p.results.name}{internal variable}
  \item{names.mixtures.indy}{internal variable}
  \item{n.sources}{internal variable}
  \item{names.sources}{internal variable}
  \item{n.isotopes}{internal variable}
  \item{names.isotopes}{internal variable}
  \item{M.actual}{internal variable}
  \item{output.mcmc.diagnostics.filename}{internal variable}
  \item{filename.prefix}{internal variable}
  \item{analysis.name}{internal variable}
  \item{plot.format.list}{internal variable}
}
\details{interal function for sisus.run()}
\value{internal variable}

\author{Erik Barry Erhardt <erik@statacumen.com>}





