\name{model.mass.balance.equation.inverse}
\alias{model.mass.balance.equation.inverse}
\title{Inverse Mass balance equations for converting biomass sampled values back to isotope values}
\description{internal function for sisus}
\usage{model.mass.balance.equation.inverse(M, n.vertices, n.isotopes, 
    n.sources, p.biomass.sam, concentration.sources, efficiency.sources, 
    biomass.per.individual.sources, number.of.individuals.sources)}
\arguments{
  \item{M}{internal variable}
  \item{n.vertices}{internal variable}
  \item{n.isotopes}{internal variable}
  \item{n.sources}{internal variable}
  \item{p.biomass.sam}{internal variable}
  \item{concentration.sources}{internal variable}
  \item{efficiency.sources}{internal variable}
  \item{biomass.per.individual.sources}{internal variable}
  \item{number.of.individuals.sources}{internal variable}
}
\details{interal function for sisus.run()}
\value{internal variable}

\author{Erik Barry Erhardt <erik@statacumen.com>}





