\name{model.mass.balance.equation}
\alias{model.mass.balance.equation}
\title{Mass balance equations for converting isotope values to biomass values}
\description{internal function for sisus}
\usage{model.mass.balance.equation(isotopes.mixtures, isotopes.sources, 
    concentration.sources, efficiency.sources, biomass.per.individual.sources, 
    number.of.individuals.sources)}
\arguments{
  \item{isotopes.mixtures}{internal variable}
  \item{isotopes.sources}{internal variable}
  \item{concentration.sources}{internal variable}
  \item{efficiency.sources}{internal variable}
  \item{biomass.per.individual.sources}{internal variable}
  \item{number.of.individuals.sources}{internal variable}
}
\details{interal function for sisus.run()}
\value{internal variable}

\author{Erik Barry Erhardt <erik@statacumen.com>}





