\name{numerical.summaries}
\alias{numerical.summaries}
\title{Numerical summaries}
\description{internal function for sisus}
\usage{numerical.summaries(p.results, p.results.name, n.sources, names.sources, 
    names.mixtures.indy, M.actual, n.vertices, analysis.name, 
    sample.name, output.stats.filename, mixtures.group.name, 
    output.stats.timeseries.filename, output.stats.timeseries.filename.csv, 
    i.mixture, SW)}
\arguments{
  \item{p.results}{internal variable}
  \item{p.results.name}{internal variable}
  \item{n.sources}{internal variable}
  \item{names.sources}{internal variable}
  \item{names.mixtures.indy}{internal variable}
  \item{M.actual}{internal variable}
  \item{n.vertices}{internal variable}
  \item{analysis.name}{internal variable}
  \item{sample.name}{internal variable}
  \item{output.stats.filename}{internal variable}
  \item{mixtures.group.name}{internal variable}
  \item{output.stats.timeseries.filename}{internal variable}
  \item{output.stats.timeseries.filename.csv}{internal variable}
  \item{i.mixture}{internal variable}
  \item{SW}{internal variable}
}
\details{interal function for sisus.run()}
\value{internal variable}

\author{Erik Barry Erhardt <erik@statacumen.com> [aut, cre]}





