\name{polytope.constraints}
\alias{polytope.constraints}
\title{Polytope constraints (incorporates both Isotope and Additional linear constraints)}
\description{internal function for sisus}
\usage{polytope.constraints(lc, n.sources, n.isotopes, biomass.values, 
    simplex.include.sw)}
\arguments{
  \item{lc}{internal variable}
  \item{n.sources}{internal variable}
  \item{n.isotopes}{internal variable}
  \item{biomass.values}{internal variable}
  \item{simplex.include.sw}{internal variable}
}
\details{interal function for sisus.run()}
\value{internal variable}

\author{Erik Barry Erhardt <erik@statacumen.com>}





