\name{polytope.multiple.samples}
\alias{polytope.multiple.samples}
\title{Defines linear constraints for each sample from the isotope distribution, passed to sample.from.polytope()}
\description{internal function for sisus}
\usage{polytope.multiple.samples(M, n.samples.isotope.mvn, n.isotopes, 
    n.sources, isotope.mvn.sample.indy.mixture, concentration.sources, 
    efficiency.sources, biomass.per.individual.sources, number.of.individuals.sources, 
    lc, skip, burnin, names.sources, SW)}
\arguments{
  \item{M}{internal variable}
  \item{n.samples.isotope.mvn}{internal variable}
  \item{n.isotopes}{internal variable}
  \item{n.sources}{internal variable}
  \item{isotope.mvn.sample.indy.mixture}{internal variable}
  \item{concentration.sources}{internal variable}
  \item{efficiency.sources}{internal variable}
  \item{biomass.per.individual.sources}{internal variable}
  \item{number.of.individuals.sources}{internal variable}
  \item{lc}{internal variable}
  \item{skip}{internal variable}
  \item{burnin}{internal variable}
  \item{names.sources}{internal variable}
  \item{SW}{internal variable}
}
\details{interal function for sisus.run()}
\value{internal variable}

\author{Erik Barry Erhardt <erik@statacumen.com> [aut, cre]}





