\name{prior.on.p}
\alias{prior.on.p}
\title{Prior on the vector of source proportions, p, is defined by the given vector alpha}
\description{internal function for sisus}
\usage{prior.on.p(priors.sources, priors.precision, n.sources)}
\arguments{
  \item{priors.sources}{internal variable}
  \item{priors.precision}{internal variable}
  \item{n.sources}{internal variable}
}
\details{interal function for sisus.run()}
\value{internal variable}

\author{Erik Barry Erhardt <erik@statacumen.com> [aut, cre]}





