\name{progress.time}
\alias{progress.time}
\title{returns the time taken since the time started}
\description{internal function for sisus}
\usage{progress.time(time.start)}
\arguments{
  \item{time.start}{internal variable}
}
\details{interal function for sisus.run()}
\value{internal variable}

\author{Erik Barry Erhardt <erik@statacumen.com> [aut, cre]}





