\name{resample.dirichlet.p}
\alias{resample.dirichlet.p}
\title{Resamples weighted samples from the polytope from the Dirichlet distribution on the p vector}
\description{internal function for sisus}
\usage{resample.dirichlet.p(M.re, sam.re, alpha.p.dirichlet, n.vertices)}
\arguments{
  \item{M.re}{internal variable}
  \item{sam.re}{internal variable}
  \item{alpha.p.dirichlet}{internal variable}
  \item{n.vertices}{internal variable}
}
\details{interal function for sisus.run()}
\value{internal variable}

\author{Erik Barry Erhardt <erik@statacumen.com>}





