\name{s.plot.convex.hull}
\alias{s.plot.convex.hull}
\title{Convex Hull plot}
\description{internal function for sisus}
\usage{s.plot.convex.hull(ch.isotopes.sources, ch.isotopes.mixtures, 
    ch.n.mixtures, n.sources, n.isotopes, names.sources, names.isotopes, 
    ch.names.mixtures, title.mixture, ch.isotope.mvn.sample, 
    ch.concentration.sources, ch.efficiency.sources, ch.isotope.sigma, 
    tit, analysis.name, filename.prefix, ch.plot.filename, plot.format.list, 
    ch.matrix.sw = 0)}
\arguments{
  \item{ch.isotopes.sources}{internal variable}
  \item{ch.isotopes.mixtures}{internal variable}
  \item{ch.n.mixtures}{internal variable}
  \item{n.sources}{internal variable}
  \item{n.isotopes}{internal variable}
  \item{names.sources}{internal variable}
  \item{names.isotopes}{internal variable}
  \item{ch.names.mixtures}{internal variable}
  \item{title.mixture}{internal variable}
  \item{ch.isotope.mvn.sample}{internal variable}
  \item{ch.concentration.sources}{internal variable}
  \item{ch.efficiency.sources}{internal variable}
  \item{ch.isotope.sigma}{internal variable}
  \item{tit}{internal variable}
  \item{analysis.name}{internal variable}
  \item{filename.prefix}{internal variable}
  \item{ch.plot.filename}{internal variable}
  \item{plot.format.list}{internal variable}
  \item{ch.matrix.sw}{internal variable}
}
\details{interal function for sisus.run()}
\value{internal variable}

\author{Erik Barry Erhardt <erik@statacumen.com>}





