\name{s.plot.convex.hull.titles}
\alias{s.plot.convex.hull.titles}
\title{outside titles Convex Hull matrix plot}
\description{internal function for sisus}
\usage{s.plot.convex.hull.titles(n.sources, n.isotopes, names.isotopes, 
    analysis.name, title.mixture, tit)}
\arguments{
  \item{n.sources}{internal variable}
  \item{n.isotopes}{internal variable}
  \item{names.isotopes}{internal variable}
  \item{analysis.name}{internal variable}
  \item{title.mixture}{internal variable}
  \item{tit}{internal variable}
}
\details{interal function for sisus.run()}
\value{internal variable}

\author{Erik Barry Erhardt <erik@statacumen.com> [aut, cre]}





