\name{s.plot.marginal.histogram}
\alias{s.plot.marginal.histogram}
\title{Plot marginal histograms}
\description{internal function for sisus}
\usage{s.plot.marginal.histogram(p.results, p.results.name, n.sources, 
    n.isotopes, M.actual, hist.bins, names.isotopes, names.mixtures.indy, 
    names.sources, analysis.name, filename.prefix, plot.filename, 
    plot.format.list, SW)}
\arguments{
  \item{p.results}{internal variable}
  \item{p.results.name}{internal variable}
  \item{n.sources}{internal variable}
  \item{n.isotopes}{internal variable}
  \item{M.actual}{internal variable}
  \item{hist.bins}{internal variable}
  \item{names.isotopes}{internal variable}
  \item{names.mixtures.indy}{internal variable}
  \item{names.sources}{internal variable}
  \item{analysis.name}{internal variable}
  \item{filename.prefix}{internal variable}
  \item{plot.filename}{internal variable}
  \item{plot.format.list}{internal variable}
  \item{SW}{internal variable}
}
\details{interal function for sisus.run()}
\value{internal variable}

\author{Erik Barry Erhardt <erik@statacumen.com>}





