\name{s.plot.settings.begin.end}
\alias{s.plot.settings.begin.end}
\title{Plot settings for all plots}
\description{internal function for sisus}
\usage{s.plot.settings.begin.end(filename.prefix = "SISUS", plot.filename = "default_filename", 
    plot.mode = "settings", plot.format = 1)}
\arguments{
  \item{filename.prefix}{internal variable}
  \item{plot.filename}{internal variable}
  \item{plot.mode}{internal variable}
  \item{plot.format}{internal variable}
}
\details{interal function for sisus.run()}
\value{internal variable}

\author{Erik Barry Erhardt <erik@statacumen.com> [aut, cre]}





