\name{sample.from.polytope}
\alias{sample.from.polytope}
\title{CORE function to draw the samples from the polytope}
\description{internal function for sisus}
\usage{sample.from.polytope(Ab, M, skip, burnin, warning.sw, i.samples.isotope.mvn = 1)}
\arguments{
  \item{Ab}{internal variable}
  \item{M}{internal variable}
  \item{skip}{internal variable}
  \item{burnin}{internal variable}
  \item{warning.sw}{internal variable}
  \item{i.samples.isotope.mvn}{internal variable}
}
\details{interal function for sisus.run()}
\value{internal variable}

\author{Erik Barry Erhardt <erik@statacumen.com> [aut, cre]}





