\name{sisus-package}
\alias{sisus-package}
\alias{sisus}
\docType{package}
\title{SISUS: Stable Isotope Sourcing using Sampling}
\description{SISUS for source partitioning using stable isotopes.
SISUS reads in a specific Excel-like workbook and performs an IsoSource-type analysis by
returning a sample of feasible solutions to p in the non-over constrained linear systems, b=Ap.
Edit \file{sisus_*_template.xls} and input data values and run parameters.
Run \code{\link{sisus.run}(filename)}. See output in current working directory.}
\details{
\tabular{ll}{Package: \tab sisus\cr
Version: \tab 3.9-12\cr
Date: \tab 2014-03-10\cr
Title: \tab SISUS: Stable Isotope Sourcing using Sampling\cr
Authors@R: \tab person(given = "Erik Barry", family = "Erhardt", email = "erik@statacumen.com", role = c("aut", "cre"))\cr
Depends: \tab R (>= 3.0.0)\cr
Imports: \tab MASS, RColorBrewer, coda, gtools, moments, polyapost, rcdd, stats\cr
Suggests: \tab gdata\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
URL: \tab http://StatAcumen.com/sisus\cr
Author: \tab Erik Barry Erhardt <erik@statacumen.com> [aut, cre]\cr
Maintainer: \tab Erik Barry Erhardt <erik@statacumen.com> [aut, cre]\cr}
}
\author{Erik Barry Erhardt <erik@statacumen.com> [aut, cre]}

\keyword{ package }


