\name{write.Ab}
\alias{write.Ab}
\title{write A matrix and b vector}
\description{writes out the linear constraints to the process_info.txt file}
\usage{write.Ab(Ab)}
\arguments{
  \item{Ab}{internal variable}
}
\details{interal function for sisus.run()}
\value{internal variable}

\author{Erik Barry Erhardt <erik@statacumen.com> [aut, cre]}





