\name{write.input}
\alias{write.input}
\title{Write input values out to a file}
\description{internal function for sisus}
\usage{write.input(DATA, filename, output.inputs.filename)}
\arguments{
  \item{DATA}{internal variable}
  \item{filename}{internal variable}
  \item{output.inputs.filename}{internal variable}
}
\details{interal function for sisus.run()}
\value{internal variable}

\author{Erik Barry Erhardt <erik@statacumen.com>}





