\name{write.model.settings}
\alias{write.model.settings}
\title{Write model settings to the process log}
\description{internal function for sisus}
\usage{write.model.settings(n.mixtures, n.sources, n.isotopes, analysis.name, 
    filename.prefix, mixtures.group.name, SW)}
\arguments{
  \item{n.mixtures}{internal variable}
  \item{n.sources}{internal variable}
  \item{n.isotopes}{internal variable}
  \item{analysis.name}{internal variable}
  \item{filename.prefix}{internal variable}
  \item{mixtures.group.name}{internal variable}
  \item{SW}{internal variable}
}
\details{interal function for sisus.run()}
\value{internal variable}

\author{Erik Barry Erhardt <erik@statacumen.com>}





