\name{write.out}
\alias{write.out}
\title{write directly to the screen}
\description{internal function for sisus}
\usage{write.out(text.to.cat)}
\arguments{
  \item{text.to.cat}{internal variable}
}
\details{interal function for sisus.run()}
\value{internal variable}

\author{Erik Barry Erhardt <erik@statacumen.com> [aut, cre]}





