\name{write.progress}
\alias{write.progress}
\title{write progress to the screen}
\description{internal function for sisus}
\usage{write.progress(text.to.cat, time.start)}
\arguments{
  \item{text.to.cat}{internal variable}
  \item{time.start}{internal variable}
}
\details{interal function for sisus.run()}
\value{internal variable}

\author{Erik Barry Erhardt <erik@statacumen.com> [aut, cre]}





