\name{sisus-package}
\alias{sisus-package}
\alias{sisus-package}
\docType{package}
\title{
sisus reads in a specific Excel-like workbook and performs an IsoSource-type analysis
}
\description{
Provides sample of feasible solutions to pi in linear systems, b = A*pi. \cr
Designed in the language of stable isotope mixing models for ecological and biological applications.
}
\details{
\tabular{ll}{
Package: \tab sisus\cr
Type: \tab Package\cr
Version: \tab 0.09-011\cr
Date: \tab 2008-05-22\cr
License: \tab GNU General Public License (GPL)\cr
}
sisus(filename) is the only user function
}
\author{
Erik Barry Erhardt

Maintainer: Erik Barry Erhardt <erik@statacumen.com>
}
%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
%\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
\examples{
## # set working directory for many output files with setwd()
## # see http://statacumen.com/sisus for workbook
## filename = "http://statacumen.com/old/sisus/examples/SISUS_v0_09_template.xls";
## sisus(filename)
}
